More sample ships
Another batch of sample ships, and a pretty nice template for doing the
tests. The "cost" tests are currently being skipped, because they're
definitely not ready. The validation tests are really just coal-mine
canaries -- these should all be valid, so if an error shows up, that
means I've got something wrong.