Printable version of a ship
This is loosely based on the stat blocks in the rulebook. It's (more or
less) human readable, and it's pretty easily parsable, so it can serve
as the serialization format, much like the SoBH builder.

Testing this also sort of serves as an integration test on whether or
not the stats pick up the bonuses from upgrades.